requirements.txt
django
django-allauth
django-environ
django-split-settings

python-webpack-boilerplate

# We are only using this for helpers. Bootstrap js/css comes from webpack
django-bootstrap-v5
